Moving memory store to test module; Removing not-really-useful convenience method.
